move context event to events
